Fix : Delete - uint8

+++ tests/detection/decoding.py
@@ -6,7 +6,7 @@ import numpy as np
 
 def fun_decoding(region_score_map):
     region_score_map = np.around(region_score_map)
-    region_score_map = region_score_map.astype("uint8")
+    # region_score_map = region_score_map.astype("uint8")
     cv2.imshow("img", region_score_map)
     cv2.waitKey(0)
     cv2.destroyAllWindows()
